requirements.txt
pandas
python-pptx
fpdf
matplotlib
google-generativeai
python-dotenv
sqlalchemy
openpyxl
streamlit
reportlab
seaborn

